feat: create dockerfile move go files around to follow best practices
